flood_dataset/flood-detection-cnn/requirements.txt
numpy
torch
torchvision
rasterio
opencv-python
scikit-learn
matplotlib
tqdm
PyYAML
albumentations
seaborn